feat: Create livekit app
