+++ totem/circles/mobile_api.py
@@ -16,11 +16,9 @@ from totem.circles.filters import (
     space_detail_schema,
     upcoming_recommended_events,
 )
-from totem.circles.livekit import livekit_create_access_token
 from totem.circles.models import Circle, CircleEvent, CircleEventException
 from totem.circles.schemas import (
     EventDetailSchema,
-    LivekitTokenResponseSchema,
     SpaceSchema,
     SummarySpacesSchema,
 )
@@ -234,26 +232,3 @@ def rsvp_cancel(request: HttpRequest, event_slug: str):
     except CircleEventException as e:
         raise AuthorizationError(message=str(e))
     return True
-
-
-@spaces_router.get(
-    "/event/{event_slug}/token",
-    response={200: LivekitTokenResponseSchema, 403: str, 404: str},
-    tags=["spaces"],
-    url_name="livekit_token",
-)
-def get_livekit_token(request, event_slug: str):
-    user: User = request.auth
-    event = get_object_or_404(CircleEvent, slug=event_slug)
-
-    if not event.joinable:
-        raise AuthorizationError(message="Event is not joinable at this time.")
-
-    if not event.attendees.filter(id=user.id).exists():
-        raise AuthorizationError(message="You have not RSVP'd for this event.")
-
-    try:
-        token = livekit_create_access_token(user, event)
-        return LivekitTokenResponseSchema(token=token)
-    except ValueError as e:
-        raise AuthorizationError(message=str(e))
