+++ totem/livekit/schemas.py
@@ -1,0 +1,5 @@
+from ninja import Schema
+
+
+class LivekitTokenResponseSchema(Schema):
+    token: str
